feat(core): share backup so every client can export
